email notification Back-end commit

+++ apps/workspaces/tasks.py
@@ -1,5 +1,12 @@
-from datetime import datetime
+import time
+from datetime import datetime, timedelta, date
+from typing import List
 
+from django.conf import settings
+from django.db.models import Q
+
+from django.core.mail import EmailMessage
+from django.template.loader import render_to_string
 from django_q.models import Schedule
 
 from apps.fyle.models import ExpenseGroup
@@ -7,18 +14,46 @@ from apps.fyle.tasks import async_create_expense_groups
 from apps.quickbooks_online.tasks import schedule_bills_creation, schedule_cheques_creation, \
     schedule_journal_entry_creation, schedule_credit_card_purchase_creation, schedule_qbo_expense_creation
 from apps.tasks.models import TaskLog
-from apps.workspaces.models import WorkspaceSchedule, WorkspaceGeneralSettings, LastExportDetail
+from apps.workspaces.models import User, Workspace, WorkspaceSchedule, WorkspaceGeneralSettings, LastExportDetail, QBOCredential
+from fyle_accounting_mappings.models import MappingSetting, ExpenseAttribute
+
+def schedule_email_notification(workspace_id: int, schedule_enabled: bool, hours: int):
+    if schedule_enabled:
+        schedule, _ = Schedule.objects.update_or_create(
+            func='apps.workspaces.tasks.run_email_notification',
+            args='{}'.format(workspace_id),
+            defaults={
+                'schedule_type': Schedule.MINUTES,
+                'minutes': hours * 60,
+                'next_run': datetime.now() + timedelta(minutes=10)
+            }
+        )
+    else:
+        schedule: Schedule = Schedule.objects.filter(
+            func='apps.workspaces.tasks.run_email_notification',
+            args='{}'.format(workspace_id)
+        ).first()
+
+        if schedule:
+            schedule.delete()
 
 
-def schedule_sync(workspace_id: int, schedule_enabled: bool, hours: int):
+def schedule_sync(workspace_id: int, schedule_enabled: bool, hours: int, email_added: List, emails_selected: List):
     ws_schedule, _ = WorkspaceSchedule.objects.get_or_create(
         workspace_id=workspace_id
     )
+
+    schedule_email_notification(workspace_id=workspace_id, schedule_enabled=schedule_enabled, hours=hours)
 
     if schedule_enabled:
         ws_schedule.enabled = schedule_enabled
         ws_schedule.start_datetime = datetime.now()
         ws_schedule.interval_hours = hours
+        ws_schedule.emails_selected = emails_selected
+        
+        if email_added:
+            ws_schedule.additional_email_options.append(email_added)
+
 
         schedule, _ = Schedule.objects.update_or_create(
             func='apps.workspaces.tasks.run_sync_schedule',
@@ -29,17 +64,17 @@ def schedule_sync(workspace_id: int, schedule_enabled: bool, hours: int):
                 'next_run': datetime.now()
             }
         )
+
         ws_schedule.schedule = schedule
 
         ws_schedule.save()
 
-    elif not schedule_enabled:
+    elif not schedule_enabled and ws_schedule.schedule:
         schedule = ws_schedule.schedule
-        if schedule:
-            ws_schedule.enabled = schedule_enabled
-            ws_schedule.schedule = None
-            ws_schedule.save()
-            schedule.delete()
+        ws_schedule.enabled = schedule_enabled
+        ws_schedule.schedule = None
+        ws_schedule.save()
+        schedule.delete()
 
     return ws_schedule
 
@@ -131,3 +166,57 @@ def export_to_qbo(workspace_id, export_mode=None):
             schedule_bills_creation(
                 workspace_id=workspace_id, expense_group_ids=expense_group_ids
             )
+
+def run_email_notification(workspace_id):
+
+    ws_schedule, _ = WorkspaceSchedule.objects.get_or_create(
+        workspace_id=workspace_id
+    )
+
+    task_logs = TaskLog.objects.filter(
+        ~Q(type__in=['CREATING_REIMBURSEMENT', 'FETCHING_EXPENSES', 'CREATING_AP_PAYMENT']),
+        workspace_id=workspace_id,
+        status='FAILED'
+    )
+
+    workspace = Workspace.objects.get(id=workspace_id)
+    admin_data = WorkspaceSchedule.objects.get(workspace_id=workspace_id)
+    qbo = QBOCredential.objects.get(workspace=workspace)
+
+    if ws_schedule.enabled:
+        for admin_email in admin_data.emails_selected:
+            attribute = ExpenseAttribute.objects.filter(workspace_id=workspace_id, value=admin_email).first()
+
+            if attribute:
+                admin_name = attribute.detail['full_name']
+            else:
+                for data in admin_data.additional_email_options:
+                    if data['email'] == admin_email:
+                        admin_name = data['name']
+
+            if task_logs and (ws_schedule.error_count is None or len(task_logs) > ws_schedule.error_count):
+                context = {
+                    'name': admin_name,
+                    'errors': len(task_logs),
+                    'fyle_company': workspace.name,
+                    'qbo_company': qbo.company_name,
+                    'workspace_id': workspace_id,
+                    'export_time': workspace.last_synced_at.date(),
+                    'year': date.today().year,
+                    'app_url': "{0}/workspaces/{1}/expense_groups".format(settings.FYLE_APP_URL, workspace_id)
+                    }
+                message = render_to_string("mail_template.html", context)
+
+                mail = EmailMessage(
+                    subject="Export To QuickBooks Failed",
+                    body=message,
+                    from_email=settings.EMAIL,
+                    to=[admin_email],
+                )
+
+                mail.content_subtype = "html"
+                mail.send()
+
+        ws_schedule.error_count = len(task_logs)
+        ws_schedule.save()
+
